# howllo-server/openapi.yaml
openapi: 3.0.3
info:
  title: Howllo Server API
  description: Feedback management backend with boards, posts, roadmap, search, AI suggestions, webhooks, and API tokens.
  version: v1
  contact:
    name: Howllo Team
servers:
  - url: http://localhost:7700/api
    description: Local development
paths:
...
  /admin/members:
    get:
      summary: List members
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Member list
    post:
      summary: Add member
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Member added
  /admin/members/{user_id}/role:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Change member role
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Role changed
  /admin/members/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: string }
    delete:
      summary: Remove member
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Member removed
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Howllo account/workspace session, optional legacy RooIAM token, or scoped API token